requirements.txt
pathlib==1.0.1
langchain==0.3.2
langchain-community==0.3.1
langchain-openai==0.3.11
langchain-core==0.3.49
python-dotenv==1.0.1
faiss-cpu==1.8.0
pytest==8.3.4
streamlit==1.36.0
numpy==1.26.4
ipykernel
pypdf==5.4.0